create Home layout
