+++ src/layouts/Home/index.tsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import Home from './Home';
+
+export default Home;
