+++ src/layouts/Home/styles.ts
@@ -1,0 +1,93 @@
+import { Theme, makeStyles, createStyles } from '@material-ui/core';
+
+const drawerWidth = 280;
+
+const styles: any = (theme: Theme) =>
+    createStyles({
+        root: {
+            display: 'flex',
+        },
+        appBar: {
+            width: `calc(100% - ${drawerWidth}px)`,
+            marginLeft: drawerWidth,
+        },
+        toolbar: {
+            display: 'flex',
+            flexDirection: 'row',
+            justifyContent: 'space-between',
+            alignItems: 'center',
+        },
+        drawer: {
+            width: drawerWidth,
+            flexShrink: 0,
+        },
+        drawerPaper: {
+            width: drawerWidth,
+            backgroundColor: theme.palette.secondary.main,
+            display: 'flex',
+            flexDirection: 'column',
+            justifyContent: 'space-between',
+        },
+        // necessary for content to be below app bar
+        toolbarSpacer: theme.mixins.toolbar,
+        content: {
+            width: `calc(100% - ${drawerWidth}px)`,
+            height: '100vh',
+            maxHeight: '100vh',
+            flexGrow: 1,
+            backgroundColor: theme.palette.secondary.main,
+            display: 'flex',
+            flexDirection: 'column',
+            justifyContent: 'flex-start',
+        },
+        mainContainer: {
+            flex: 1,
+            display: 'flex',
+            flexDirection: 'column',
+            backgroundColor: theme.palette.secondary.main,
+        },
+        messagesListContainer: {
+            overflow: 'scroll',
+            flex: 1,
+            display: 'flex',
+            flexDirection: 'column',
+            justifyContent: 'flex-start',
+            alignItems: 'center',
+        },
+        messageBox: {
+            display: 'flex',
+            flexDirection: 'row',
+            justifyContent: 'space-between',
+            backgroundColor: theme.palette.primary.light,
+        },
+        drawerToolbarContent: {
+            display: 'flex',
+            flexDirection: 'column',
+            justifyContent: 'center',
+            alignItems: 'flex-start',
+            color: 'white',
+        },
+        white: {
+            color: 'white',
+        },
+        activeAddPeer: {
+            color: theme.palette.secondary.light,
+        },
+        bothButtonText: {
+            color: theme.palette.primary.main,
+            fontSize: '0.9rem',
+        },
+        activeCall: {
+            backgroundColor: theme.palette.system.success,
+            color: theme.palette.common.white,
+        },
+        mediaToggleIcons: {
+            fontSize: '1.1rem',
+            color: theme.palette.primary.main,
+        },
+        endCallButton: {
+            color: theme.palette.system.error,
+        },
+    });
+
+export default makeStyles(styles);
